refactor: extract cli package and markdowntoc package
