+++ .gitignore
@@ -1,1 +1,1 @@
-markdown-toc
+/markdown-toc
