+++ Makefile
@@ -1,2 +1,2 @@
 build:
-	go build -o markdown-toc ./...
+	go build -o markdown-toc ./cmd/markdown-toc
